+++ tests/test_enterprise/api/test_views.py
@@ -29,7 +29,8 @@ from testfixtures import LogCapture
 from django.conf import settings
 from django.contrib.auth.models import Permission
 from django.core.cache import cache
-from django.test import override_settings
+from django.http import HttpRequest
+from django.test import Client, TestCase, override_settings
 from django.utils import timezone
 
 from enterprise.api.v1 import serializers
@@ -72,6 +73,7 @@ from enterprise.models import (
 from enterprise.roles_api import admin_role
 from enterprise.toggles import (
     ADMIN_PORTAL_LEARNER_PROFILE_VIEW_ENABLED,
+    CATALOG_QUERY_SEARCH_FILTERS_ENABLED,
     ENTERPRISE_CUSTOMER_SUPPORT_TOOL,
     ENTERPRISE_GROUPS_V1,
     ENTERPRISE_GROUPS_V2,
@@ -104,6 +106,7 @@ from test_utils import (
 )
 from test_utils.factories import (
     FAKER,
+    EnterpriseCourseEnrollmentFactory,
     EnterpriseCustomerFactory,
     EnterpriseCustomerSsoConfigurationFactory,
     EnterpriseCustomerUserFactory,
@@ -173,6 +176,26 @@ LICENSED_ENTERPRISE_COURSE_ENROLLMENTS_REVOKE_ENDPOINT = reverse(
 EXPIRED_LICENSED_ENTERPRISE_COURSE_ENROLLMENTS_ENDPOINT = reverse(
     'licensed-enterprise-course-enrollment-bulk-licensed-enrollments-expiration'
 )
+SERIALIZED_MOCK_IN_PROGRESS_ENROLLMENT = {
+    'course_id': 'in-progress-course',
+    'is_enrollment_active': True,
+    'course_run_status': 'in_progress',
+}
+SERIALIZED_MOCK_UPCOMING_ENROLLMENT = {
+    'course_id': 'upcoming-course',
+    'is_enrollment_active': True,
+    'course_run_status': 'upcoming',
+}
+SERIALIZED_MOCK_COMPLETED_ENROLLMENT = {
+    'course_id': 'completed-course',
+    'is_enrollment_active': True,
+    'course_run_status': 'completed',
+}
+SERIALIZED_MOCK_SAVED_ENROLLMENT = {
+    'course_id': 'saved-for-later',
+    'is_enrollment_active': True,
+    'course_run_status': 'saved_for_later',
+}
 
 
 def create_mock_default_enterprise_enrollment_intention(
@@ -2027,64 +2050,65 @@ class TestEnterpriseCustomerViewSet(BaseTestEnterpriseAPIViews):
     @ddt.data(
         # Request missing required permissions query param.
         (True, False, [], {}, False, {'detail': 'User is not allowed to access the view.'},
-         False, False, False, False, False, False, False),
+         False, False, False, False, False, False, False, False),
         # Staff user that does not have the specified group permission.
         (True, False, [], {'permissions': ['enterprise_enrollment_api_access']}, False,
-         {'detail': 'User is not allowed to access the view.'}, False, False, False, False, False, False, False),
+         {'detail': 'User is not allowed to access the view.'}, False, False, False, False, False, False, False, False),
         # Staff user that does have the specified group permission.
         (True, False, ['enterprise_enrollment_api_access'], {'permissions': ['enterprise_enrollment_api_access']},
-         True, None, False, False, False, False, False, False, False),
+         True, None, False, False, False, False, False, False, False, False),
         # Non staff user that is not linked to the enterprise, nor do they have the group permission.
         (False, False, [], {'permissions': ['enterprise_enrollment_api_access']}, False,
-         {'detail': 'User is not allowed to access the view.'}, False, False, False, False, False, False, False),
+         {'detail': 'User is not allowed to access the view.'}, False, False, False, False, False, False, False, False),
         # Non staff user that is not linked to the enterprise, but does have the group permission.
         (False, False, ['enterprise_enrollment_api_access'], {'permissions': ['enterprise_enrollment_api_access']},
-         False, None, False, False, False, False, False, False, False),
+         False, None, False, False, False, False, False, False, False, False),
         # Non staff user that is linked to the enterprise, but does not have the group permission.
         (False, True, [], {'permissions': ['enterprise_enrollment_api_access']}, False,
-         {'detail': 'User is not allowed to access the view.'}, False, False, False, False, False, False, False),
+         {'detail': 'User is not allowed to access the view.'}, False, False, False, False, False, False, False, False),
         # Non staff user that is linked to the enterprise and does have the group permission
         (False, True, ['enterprise_enrollment_api_access'], {'permissions': ['enterprise_enrollment_api_access']},
-         True, None, False, False, False, False, False, False, False),
+         True, None, False, False, False, False, False, False, False, False),
         # Non staff user that is linked to the enterprise and has group permission and the request has passed
         # multiple groups to check.
         (False, True, ['enterprise_enrollment_api_access'],
          {'permissions': ['enterprise_enrollment_api_access', 'enterprise_data_api_access']}, True, None, False,
-         False, False, False, False, False, False),
+         False, False, False, False, False, False, False),
         # Staff user with group permission filtering on non existent enterprise id.
         (True, False, ['enterprise_enrollment_api_access'],
          {'permissions': ['enterprise_enrollment_api_access'], 'enterprise_id': FAKE_UUIDS[1]}, False,
-         None, False, False, False, False, False, False, False),
+         None, False, False, False, False, False, False, False, False),
         # Staff user with group permission filtering on enterprise id successfully.
         (True, False, ['enterprise_enrollment_api_access'],
          {'permissions': ['enterprise_enrollment_api_access'], 'enterprise_id': FAKE_UUIDS[0]}, True,
-         None, False, False, False, False, False, False, False),
+         None, False, False, False, False, False, False, False, False),
         # Staff user with group permission filtering on search param with no results.
         (True, False, ['enterprise_enrollment_api_access'],
          {'permissions': ['enterprise_enrollment_api_access'], 'search': 'blah'}, False,
-         None, False, False, False, False, False, False, False),
+         None, False, False, False, False, False, False, False, False),
         # Staff user with group permission filtering on search param with results.
         (True, False, ['enterprise_enrollment_api_access'],
          {'permissions': ['enterprise_enrollment_api_access'], 'search': 'test'}, True,
-         None, False, False, False, False, False, False, False),
+         None, False, False, False, False, False, False, False, False),
         # Staff user with group permission filtering on slug with results.
         (True, False, ['enterprise_enrollment_api_access'],
          {'permissions': ['enterprise_enrollment_api_access'], 'slug': TEST_SLUG}, True,
-         None, False, False, False, False, False, False, False),
+         None, False, False, False, False, False, False, False, False),
         # Staff user with group permissions filtering on slug with no results.
         (True, False, ['enterprise_enrollment_api_access'],
          {'permissions': ['enterprise_enrollment_api_access'], 'slug': 'blah'}, False,
-         None, False, False, False, False, False, False, False),
+         None, False, False, False, False, False, False, False, False),
         # Staff user with group permission filtering on slug with results, with
         # top down assignment & real-time LCM feature enabled,
         # prequery search results enabled and
         # enterprise groups v1 feature enabled
         # enterprise groups v2 feature enabled
         # enterprise customer support tool enabled
-        # enterprise learner bff enabled
+        # enterprise learner bff enabled,
+        # admin portal learner profile view enabled
         (True, False, ['enterprise_enrollment_api_access'],
          {'permissions': ['enterprise_enrollment_api_access'], 'slug': TEST_SLUG}, True,
-         None, True, True, True, True, True, True, True),
+         None, True, True, True, True, True, True, True, True),
     )
     @ddt.unpack
     @mock.patch('enterprise.utils.get_logo_url')
@@ -2103,6 +2127,7 @@ class TestEnterpriseCustomerViewSet(BaseTestEnterpriseAPIViews):
             enterprise_customer_support_tool_enabled,
             enterprise_learner_bff_enabled,
             admin_portal_learner_profile_view_enabled,
+            catalog_query_search_filters_enabled,
             mock_get_logo_url,
     ):
         """
@@ -2152,7 +2177,6 @@ class TestEnterpriseCustomerViewSet(BaseTestEnterpriseAPIViews):
             TOP_DOWN_ASSIGNMENT_REAL_TIME_LCM,
             active=is_top_down_assignment_real_time_lcm_enabled
         ):
-
             response = client.get(
                 f"{settings.TEST_SERVER}{ENTERPRISE_CUSTOMER_WITH_ACCESS_TO_ENDPOINT}?{urlencode(query_params, True)}"
             )
@@ -2160,7 +2184,6 @@ class TestEnterpriseCustomerViewSet(BaseTestEnterpriseAPIViews):
             FEATURE_PREQUERY_SEARCH_SUGGESTIONS,
             active=feature_prequery_search_suggestions_enabled
         ):
-
             response = client.get(
                 f"{settings.TEST_SERVER}{ENTERPRISE_CUSTOMER_WITH_ACCESS_TO_ENDPOINT}?{urlencode(query_params, True)}"
             )
@@ -2195,6 +2218,13 @@ class TestEnterpriseCustomerViewSet(BaseTestEnterpriseAPIViews):
         with override_waffle_flag(
             ADMIN_PORTAL_LEARNER_PROFILE_VIEW_ENABLED,
             active=admin_portal_learner_profile_view_enabled
+        ):
+            response = client.get(
+                f"{settings.TEST_SERVER}{ENTERPRISE_CUSTOMER_WITH_ACCESS_TO_ENDPOINT}?{urlencode(query_params, True)}"
+            )
+        with override_waffle_flag(
+            CATALOG_QUERY_SEARCH_FILTERS_ENABLED,
+            active=catalog_query_search_filters_enabled
         ):
             response = client.get(
                 f"{settings.TEST_SERVER}{ENTERPRISE_CUSTOMER_WITH_ACCESS_TO_ENDPOINT}?{urlencode(query_params, True)}"
@@ -2271,6 +2301,7 @@ class TestEnterpriseCustomerViewSet(BaseTestEnterpriseAPIViews):
                     'enterprise_groups_v2': enterprise_groups_v2_enabled,
                     'enterprise_learner_bff_enabled': enterprise_learner_bff_enabled,
                     'admin_portal_learner_profile_view_enabled': admin_portal_learner_profile_view_enabled,
+                    'catalog_query_search_filters_enabled': catalog_query_search_filters_enabled,
                 }
             }
             assert response in (expected_error, mock_empty_200_success_response)
@@ -9882,3 +9913,176 @@ class TestEnterpriseUser(BaseTestEnterpriseAPIViews):
 
         assert expected_json == response.json().get('results')
         assert response.json().get('count') == 1
+
+
+class EnterpriseGroupMembershipViewSetTests(BaseTestEnterpriseAPIViews):
+    """Unit tests for EnterpriseGroupMembershipViewSet."""
+
+    def setUp(self):
+        """Set up test data."""
+        super().setUp()
+        self.enterprise_customer = EnterpriseCustomerFactory()
+        self.enterprise_customer_user = EnterpriseCustomerUserFactory(
+            user_id=self.user.id, enterprise_customer=self.enterprise_customer
+        )
+        self.group_1 = EnterpriseGroupFactory(enterprise_customer=self.enterprise_customer, group_type='flex')
+        self.group_2 = EnterpriseGroupFactory(enterprise_customer=self.enterprise_customer, group_type='budget')
+        self.enterprise_customer_user = EnterpriseCustomerUserFactory(
+            user_id="123", enterprise_customer__uuid=FAKE_UUIDS[0])
+        self.membership1 = EnterpriseGroupMembershipFactory(
+            enterprise_customer_user=self.enterprise_customer_user,
+            group=self.group_1,
+        )
+        self.membership2 = EnterpriseGroupMembershipFactory(
+            enterprise_customer_user=self.enterprise_customer_user,
+            group=self.group_2,
+        )
+
+        self.url = reverse("enterprise-group-membership")
+
+    def test_missing_required_params(self):
+        """Ensure API returns 400 Bad Request if required parameters are missing."""
+        response = self.client.get(self.url)
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertIn("error", response.json())
+
+    def test_filter_by_lms_user_id_and_enterprise_uuid(self):
+        """Ensure filtering by lms_user_id and enterprise_uuid returns correct results."""
+        response = self.client.get(self.url, {"lms_user_id": "123", "enterprise_uuid": FAKE_UUIDS[0]})
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(len(response.json()["results"]), 1)
+
+    def test_no_matching_results(self):
+        """Ensure API returns empty results if no matching records exist."""
+        response = self.client.get(self.url, {"lms_user_id": "999", "enterprise_uuid": FAKE_UUIDS[0]})
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(len(response.json()["results"]), 0)
+
+    def test_pagination_applied(self):
+        """Ensure pagination applies correctly if multiple records exist."""
+        EnterpriseGroupMembershipFactory.create_batch(15, enterprise_customer_user=self.enterprise_customer_user)
+        response = self.client.get(self.url, {"lms_user_id": "123", "enterprise_uuid": FAKE_UUIDS[0]})
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertIn("count", response.json())
+        self.assertIn("next", response.json())
+        self.assertIn("results", response.json())
+
+
+class EnterpriseCourseEnrollmentAdminViewSetTest(TestCase):
+    """Unit tests for EnterpriseCourseEnrollmentAdminViewSet."""
+    class MockSerializer:
+        """ Returns obj with data property. """
+        @property
+        def data(self):
+            """ Return fake serialized data. """
+            return [
+                SERIALIZED_MOCK_IN_PROGRESS_ENROLLMENT,
+                SERIALIZED_MOCK_UPCOMING_ENROLLMENT,
+                SERIALIZED_MOCK_COMPLETED_ENROLLMENT,
+                SERIALIZED_MOCK_SAVED_ENROLLMENT,
+            ]
+
+    def setUp(self):
+        """Set up test data."""
+        super().setUp()
+        self.enterprise_customer = EnterpriseCustomerFactory()
+        self.user = UserFactory.create(is_staff=True, is_active=True)
+        self.user.set_password("QWERTY")
+        self.user.save()
+        self.enrolled_ent_customer_user = EnterpriseCustomerUserFactory.create(
+            user_id=self.user.id,
+            enterprise_customer=self.enterprise_customer,
+        )
+        self.course_run_id = 'course-v1:edX+DemoX+Demo_Course'
+        self.enterprise_enrollment = EnterpriseCourseEnrollmentFactory.create(
+            enterprise_customer_user=self.enrolled_ent_customer_user,
+            course_id=self.course_run_id,
+        )
+        self.url = reverse("enterprise-course-enrollment-admin")
+        self.client = Client()
+
+    @mock.patch.object(HttpRequest, 'get_host', return_value='example.edx.org')
+    @mock.patch('enterprise.api.v1.serializers.EnterpriseCourseEnrollmentAdminViewSerializer')
+    @mock.patch('enterprise.api.v1.views.enterprise_course_enrollment.get_course_overviews')
+    def test_returns_correct_enterprise_course_enrollments(self, mock_get_course_overviews, mock_serializer, _,):
+        """
+        Ensure the API correctly fetches enterprise course enrollments.
+        """
+        mock_get_course_overviews.return_value = {'overview_info': 'this would be a larger dict'}
+        mock_serializer.return_value = self.MockSerializer()
+        response = self.client.get(self.url,
+                                   {'lms_user_id': self.user.id,
+                                    'enterprise_uuid': self.enterprise_customer.uuid})
+        # Returns 401 if user is not an admin
+        self.assertEqual(response.status_code, 401)
+        admin_user = UserFactory.create(is_active=True, is_staff=True)
+        admin_user.set_password("123")
+        admin_user.save()
+        self.client.login(username=admin_user.username, password="123")
+        response = self.client.get(self.url,
+                                   {'lms_user_id': self.user.id,
+                                    'enterprise_uuid': self.enterprise_customer.uuid})
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.json(), {
+            'count': 4,
+            'next': None,
+            'previous': None,
+            'results': {
+                'in_progress': [SERIALIZED_MOCK_IN_PROGRESS_ENROLLMENT],
+                'upcoming': [SERIALIZED_MOCK_UPCOMING_ENROLLMENT],
+                'completed': [SERIALIZED_MOCK_COMPLETED_ENROLLMENT],
+                'saved_for_later': [SERIALIZED_MOCK_SAVED_ENROLLMENT],
+            }
+        })
+
+    @mock.patch('enterprise.api.v1.serializers.EnterpriseCourseEnrollmentAdminViewSerializer')
+    @mock.patch('enterprise.api.v1.views.enterprise_course_enrollment.get_course_overviews')
+    def test_view_unlinked_user_returns_401_not_found(self, mock_get_course_overviews, mock_serializer):
+        """
+        Ensure 401 is returned if the user does not belong to the enterprise.
+        """
+        mock_get_course_overviews.return_value = {'overview_info': 'this would be a larger dict'}
+        mock_serializer.return_value = self.MockSerializer()
+        user = UserFactory(is_staff=True, is_active=True)
+        user.save()
+        self.client.login(username=user.username, password="123")
+        response = self.client.get(self.url,
+                                   {'lms_user_id': user.id,
+                                    'enterprise_uuid': self.enterprise_customer.uuid})
+        self.assertEqual(response.status_code, 401)
+
+    @mock.patch('enterprise.api.v1.serializers.EnterpriseCourseEnrollmentAdminViewSerializer')
+    @mock.patch('enterprise.api.v1.views.enterprise_course_enrollment.get_course_overviews')
+    def test_view_returns_empty_results(self, mock_get_course_overviews, mock_serializer):
+        """
+        Ensure empty results are returned if no courses are found for a user.
+        """
+        admin_user = UserFactory.create(is_active=True, is_staff=True)
+        admin_user.set_password("123")
+        admin_user.save()
+        mock_get_course_overviews.return_value = {}
+        mock_serializer.return_value.data = [{}]
+        no_enrollments_user = UserFactory.create(is_active=True)
+        no_enrollments_user.save()
+        EnterpriseCustomerUserFactory.create(
+            user_id=no_enrollments_user.id,
+            enterprise_customer=self.enterprise_customer,
+        )
+        self.client.login(username=admin_user.username, password="123")
+        response = self.client.get(self.url,
+                                   {'lms_user_id': no_enrollments_user.id,
+                                    'enterprise_uuid': self.enterprise_customer.uuid})
+        self.assertEqual(response.json()['results'],
+                         {'in_progress': [], 'upcoming': [], 'completed': [], 'saved_for_later': []})
+
+    def test_missing_parameters_cause_400(self):
+        """
+        Ensure an error is returned if required parameters are missing.
+        """
+        admin_user = UserFactory.create(is_active=True, is_staff=True)
+        admin_user.set_password("123")
+        admin_user.save()
+        self.client.login(username=admin_user.username, password="123")
+        response = self.client.get(self.url, {"lms_user_id": "test_user"})
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertIn("error", response.data)